Made isEmpty a Utiity
